+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+
+# 0.1.3
+- Added sleep 2 minutes at first install for API calls to respond
+- Added recipe oauth_settings to create configurations via API calls
+
 # 0.1.2
 - Added systemd support